+++ src/Console/Concerns/ReplacesModel.php
@@ -1,0 +1,50 @@
+<?php
+/*
+ * Copyright 2024 Cloud Creativity Limited
+ *
+ * Use of this source code is governed by an MIT-style
+ * license that can be found in the LICENSE file or at
+ * https://opensource.org/licenses/MIT.
+ */
+
+declare(strict_types=1);
+
+namespace LaravelJsonApi\Laravel\Console\Concerns;
+
+use Illuminate\Support\Str;
+use function array_keys;
+use function array_values;
+use function str_replace;
+
+trait ReplacesModel
+{
+
+    /**
+     * @param string $stub
+     * @param string $model
+     * @return string
+     */
+    protected function replaceModel(string $stub, string $model): string
+    {
+        $model = str_replace('/', '\\', $model);
+
+        if (Str::startsWith($model, '\\')) {
+            $namespacedModel = trim($model, '\\');
+        } else {
+            $namespacedModel = $this->qualifyModel($model);
+        }
+
+        $model = class_basename($model);
+
+        $replace = [
+            '{{ namespacedModel }}' => $namespacedModel,
+            '{{namespacedModel}}' => $namespacedModel,
+            '{{ model }}' => $model,
+            '{{model}}' => $model,
+        ];
+
+        return str_replace(
+            array_keys($replace), array_values($replace), $stub
+        );
+    }
+}