+++ src/Console/MakeSchema.php
@@ -1,25 +1,17 @@
 <?php
 /*
- * Copyright 2023 Cloud Creativity Limited
+ * Copyright 2024 Cloud Creativity Limited
  *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+ * Use of this source code is governed by an MIT-style
+ * license that can be found in the LICENSE file or at
+ * https://opensource.org/licenses/MIT.
  */
 
 declare(strict_types=1);
 
 namespace LaravelJsonApi\Laravel\Console;
 
-use Illuminate\Support\Str;
+use LaravelJsonApi\Laravel\Console\Concerns\ReplacesModel;
 use Symfony\Component\Console\Input\InputOption;
 use function array_keys;
 use function array_values;
@@ -27,6 +19,7 @@ use function str_replace;
 
 class MakeSchema extends GeneratorCommand
 {
+    use ReplacesModel;
 
     /**
      * @var string
@@ -47,7 +40,7 @@ class MakeSchema extends GeneratorCommand
      * @var string
      */
     protected $classType = 'Schema';
-
+    
     /**
      * @inheritDoc
      */
@@ -65,7 +58,7 @@ class MakeSchema extends GeneratorCommand
      */
     protected function buildClass($name)
     {
-        $stub = parent::buildClass($name);
+        $stub = $this->replaceSchema(parent::buildClass($name));
 
         $model = $this->option('model') ?: $this->guessModel();
 
@@ -77,24 +70,11 @@ class MakeSchema extends GeneratorCommand
      * @param string $model
      * @return string
      */
-    protected function replaceModel(string $stub, string $model): string
+    protected function replaceSchema(string $stub): string
     {
-        $model = str_replace('/', '\\', $model);
-
-        if (Str::startsWith($model, '\\')) {
-            $namespacedModel = trim($model, '\\');
-        } else {
-            $namespacedModel = $this->qualifyModel($model);
-        }
-
-        $model = class_basename($model);
         $schema = $this->option('proxy') ? 'ProxySchema' : 'Schema';
 
         $replace = [
-            '{{ namespacedModel }}' => $namespacedModel,
-            '{{namespacedModel}}' => $namespacedModel,
-            '{{ model }}' => $model,
-            '{{model}}' => $model,
             '{{ schema }}' => $schema,
             '{{schema}}' => $schema,
         ];
@@ -117,5 +97,4 @@ class MakeSchema extends GeneratorCommand
             ['server', 's', InputOption::VALUE_REQUIRED, 'The JSON:API server the schema exists in.'],
         ];
     }
-
 }