+++ src/Routing/RelationshipRegistrar.php
@@ -1,18 +1,10 @@
 <?php
 /*
- * Copyright 2023 Cloud Creativity Limited
+ * Copyright 2024 Cloud Creativity Limited
  *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+ * Use of this source code is governed by an MIT-style
+ * license that can be found in the LICENSE file or at
+ * https://opensource.org/licenses/MIT.
  */
 
 declare(strict_types=1);
@@ -22,6 +14,7 @@ namespace LaravelJsonApi\Laravel\Routing;
 use Illuminate\Contracts\Routing\Registrar as RegistrarContract;
 use Illuminate\Routing\Route as IlluminateRoute;
 use Illuminate\Routing\RouteCollection;
+use Illuminate\Support\Arr;
 use LaravelJsonApi\Contracts\Schema\Schema;
 use LaravelJsonApi\Core\Support\Str;
 
@@ -272,9 +265,10 @@ class RelationshipRegistrar
         $name = $this->getRelationRouteName($method, $defaultName, $options);
 
         $action = ['as' => $name, 'uses' => $this->controller.'@'.$method];
+        $middleware = $this->getMiddleware($method, $options);
 
-        if (isset($options['middleware'])) {
-            $action['middleware'] = $options['middleware'];
+        if (!empty($middleware)) {
+            $action['middleware'] = $middleware;
         }
 
         if (isset($options['excluded_middleware'])) {
@@ -282,6 +276,22 @@ class RelationshipRegistrar
         }
 
         return $action;
+    }
+
+    /**
+     * @param string $action
+     * @param array $options
+     * @return array
+     */
+    private function getMiddleware(string $action, array $options): array
+    {
+        $all = $options['middleware'] ?? [];
+        $actions = $options['action_middleware'] ?? [];
+
+        return [
+            ...$all,
+            ...Arr::wrap($actions[$action] ?? null),
+        ];
     }
 
     /**